requirements.txt
websockets
git+https://github.com/57ur14/pefile-extract-icon.git
pywin32
requests
pystray
pyinstaller
watchdog
PyQt5
win10toast
pyjwt
